Unused code cleanup
